Fixed the following issue through an update of the XSLT that generates profile schemas from the component specification:
"The XSLT adds an xml:lang attribute to the element if the 'Multilingual' 
attribute is present in the component specification (regardless of its 
value). However, the Component Registry has been outputting 
'Multilingual="false"' for non-multilingual elements for a long time. 
Therefore, a lot of profiles have this attribute all over the place. 
That is not really a bad thing, it is probably good to have this 
explicit in the specification. Except now all those elements do take an 
optional language attribute, which is not right."